+++ base_aux/base_argskwargs/novalue.py
@@ -3,7 +3,7 @@ class NoValue:
     """
     DEPRECATE???
     ---------
-    use direct ArgsEmpty???/ArgsKwargs()
+    use direct ArgsEmpty???/InitArgsKwargs()
 
     GOAL
     ----